Brought the tracing utility into its final state
I realized the the finalizer function is not interesting to have. The
order of unloading of the .so file is not well defined. As a result, it
is extremely difficult (if not impossible) to reliable use the memory
allocated inside it or use its objects at program shutdown.

Solution: Register the finalizer function inside the shared object file,
either via a destructor or with std::atexit or thirdly via
__attribute__((destructor))